mcts-blts: Add USB test asset
This USB test asset is designed to perform simultaneous transfers using
multiple endpoints. Bulk, interrupt, and isochronous transfers are
supported, with options for timing and buffering depending on type.

Architecturally the test consists of a pair of kernel modules on each end
of the USB connection, with most of the logic in an userspace test program.
The kernel modules should build against at least Linux 2.6.23 - 2.6.35.

Signed-off-by: Jussi Saavalainen <jussi.saavalainen@digia.com>
Signed-off-by: Kari Laakso <kari.laakso@digia.com>