mcts-blts: Add Video for Linux 2 test asset
The test cases in this asset include functional and performance tests for
V4L2-connected cameras. Both still and streaming video cases are included.
The test parameters are highly platform-dependent, so some effort is needed
to support additional devices.

Note that this test asset needs a separate pixel format conversion library
to build, to be published soon.

Signed-off-by: Jussi Saavalainen <jussi.saavalainen@digia.com>
Signed-off-by: Kari Laakso <kari.laakso@digia.com>