blts-ofono: Cleanup and refactor expected values in call meter cases
The use of 'expected' in check_call_meter_value was quite wrong due to it
getting called with variations of (void*)-1, char* strings, and the
occasional integer 0.

Replace this with a still-ugly GValue combo, and clean up the various
ato*() calls in the process. Also use proper data types for call meter
max value and PPU parameters.

Also (*2), do proper conversion for timeout param reading. (Why is this
a string, anyway?)

Signed-off-by: Jussi Saavalainen <jussi.saavalainen@digia.com>